requirements.txt
python-telegram-bot[job-queue]==21.5
curl_cffi==0.7.4
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pytz==2024.1
apscheduler==3.10.4
PyPDF2==3.0.1
lxml==5.1.0
